Extract feature flag catalog and add parity tests
Moved FeatureFlagDescriptor/FeatureFlagCatalog out of FeatureFlagsSectionViewModel into a dedicated catalog file, keeping UI metadata centralized and reusable. Added tests that enforce one-to-one parity between FeatureFlagConstants and FeatureFlagCatalog entries and prevent duplicate catalog flags, so missing or mismatched feature-flag metadata fails fast. Also clarified in FeatureFlagConstants docs where user-facing flag text is defined.
